Allow specifying doctype in registry

+++ src/worker_work.jl
@@ -4,7 +4,7 @@ using JSON
 
 include("DocumentationGenerator.jl")
 
-const GIT_TOKEN_FILE = if isfile(joinpath("/config/sync", "gh_auth.txt"))
+const GIT_TOKEN_FILE = if isfile(joinpath("config", "sync", "gh_auth.txt"))
     joinpath("/config/sync", "gh_auth.txt")
 else
     joinpath(@__DIR__, "gh_auth.txt")
@@ -26,10 +26,36 @@ function license(path::String, confidence=85)
     end
 end
 
-function create_docs(pspec::Pkg.Types.PackageSpec, buildpath)
+function create_docs(pspec::Pkg.Types.PackageSpec, buildpath, registry)
     _module, rootdir = DocumentationGenerator.install_and_use(pspec)
     pkgname = pspec.name
 
+    type, uri = DocumentationGenerator.get_method_from_registry(pspec, registry, rootdir)
+
+    @info "$(pkgname) specifies docs of type $(type)"
+    out = try
+        if type == "git-repo"
+            @info("building `git-repo` docs")
+            build_git_docs(pkgname, rootdir, buildpath, uri)
+        elseif type == "hosted"
+            @info("building `hosted` docs")
+            build_hosted_docs(pkgname, rootdir, buildpath, uri)
+        elseif type == "vendored"
+            @info("building `vendored` docs")
+            build_local_dir_docs(pkgname, _module, rootdir, buildpath, uri)
+        else
+            @error("invalid doctype: $type")
+            :nope, ""
+        end
+    catch err
+        @error("Error building docs of type `$type`:", exception = err)
+        errored = true
+    end
+
+    return out
+end
+
+function build_local_dir_docs(pkgname, _module, rootdir, buildpath, uri)
     # package doesn't load, so let's only use the README
     if _module === nothing
         return mktempdir() do root
@@ -41,7 +67,7 @@ function create_docs(pspec::Pkg.Types.PackageSpec, buildpath)
 
     # actual Documenter docs
     try
-        for docdir in joinpath.(rootdir, ("docs", "doc"))
+        for docdir in (uri, joinpath.(rootdir, ("docs", "doc"))...)
             if isdir(docdir)
                 makefile = joinpath(docdir, "make.jl")
                 # create customized makefile with removed deploydocs + modified makedocs
@@ -79,7 +105,50 @@ function contributor_user(dict)
     )
 end
 
-function package_docs(uuid, name, url, version, buildpath)
+function build_hosted_docs(pkgname, rootdir, buildpath, uri)
+    # js redirect
+    open(joinpath(buildpath, "index.html"), "w") do io
+        println(io,
+            """
+            <!DOCTYPE html>
+            <html>
+                <head>
+                    <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
+                    <script type="text/javascript">
+                        window.onload = function () {
+                            window.location.replace("$(uri)");
+                        }
+                    </script>
+                </head>
+                <body>
+                    Redirecting to <a href="$(uri)">$(uri)</a>.
+                </body>
+            </html>
+            """
+        )
+    end
+    # download search index
+    try
+        download(uri*"/search_index.js", joinpath(buildpath, "search_index.js"))
+    catch err
+        @error("search index download failed for $(uri)", exception = err)
+    end
+    return :real, rootdir
+end
+
+function build_git_docs(pkgname, rootdir, buildpath, uri)
+    mktempdir() do dir
+        cd(dir)
+        run(`git clone --depth=1 $(uri) docsource`)
+        docsproject = joinpath(dir, "docsource")
+        cd(docsproject)
+        build_local_dir_docs(pkgname, true, docsproject, buildpath, "")
+    end
+
+    return :real, rootdir
+end
+
+function package_docs(uuid, name, url, version, buildpath, registry)
     pspec = PackageSpec(uuid = uuid, name = name, version = version)
     @info("Generating docs for $name")
     doctype = :default
@@ -102,7 +171,7 @@ function package_docs(uuid, name, url, version, buildpath)
                 end
             else
                 Pkg.activate(envdir)
-                doctype, rootdir = create_docs(pspec, buildpath)
+                doctype, rootdir = create_docs(pspec, buildpath, registry)
                 meta["doctype"] = string(doctype)
                 meta["installs"] = true
                 @info("Done generating docs for $name")
@@ -111,21 +180,24 @@ function package_docs(uuid, name, url, version, buildpath)
              monkeypatchdocsearch(uuid, name, buildpath)
         end
     catch e
-        @error("Package $name didn't build", error = e)
+        @error("Package $name didn't build", error = e, stacktrace=stacktrace(catch_backtrace()))
         meta["installs"] = false
     end
 
     return meta
 end
+
 function monkeypatchdocsearch(uuid, name, buildpath)
     if !(get(ENV, "DISABLE_CENTRALIZED_SEARCH", false) in ("true", "1", 1))
-        @info "monkey patching search.js for $(name)"
         searchjs = joinpath(buildpath, "assets", "search.js")
-        rm(searchjs, force=true)
-        template = String(read(joinpath(@__DIR__, "search.js.template")))
-        template = replace(template, "{{{UUID}}}" => String(uuid))
-        open(searchjs, "w") do io
-            print(io, template)
+        if isfile(searchjs)
+            @info "monkey patching search.js for $(name)"
+            rm(searchjs, force=true)
+            template = String(read(joinpath(@__DIR__, "search.js.template")))
+            template = replace(template, "{{{UUID}}}" => String(uuid))
+            open(searchjs, "w") do io
+                print(io, template)
+            end
         end
     end
 end
@@ -178,8 +250,8 @@ function package_source(uuid, name, rootdir, buildpath)
     @info("Done copying source code for $(name)-$(uuid)")
 end
 
-function build(uuid, name, url, version, buildpath)
-    meta = package_docs(uuid, name, url, version, buildpath)
+function build(uuid, name, url, version, buildpath, registry)
+    meta = package_docs(uuid, name, url, version, buildpath, registry)
     merge!(meta, package_metadata(uuid, name, url, version, buildpath))
     @info "making buildpath"
     isdir(buildpath) || mkpath(joinpath(buildpath))
